python/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
Django==4.2.17
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
python-dotenv==1.0.1
redis==5.2.1
redis-cli==1.0.1
sqlparse==0.5.3
typing_extensions==4.12.2
